requirements.txt
astroid==1.4.9
isort==4.2.5
lazy-object-proxy==1.2.2
limeclient==1.0.3
mccabe==0.6.1
pylint==1.6.5
PyYAML==3.12
requests==2.13.0
six==1.10.0
untangle==1.1.0
wrapt==1.10.8
xmltodict==0.10.2
